testParking-master/pyproject.toml
[tool.poetry]
name = "testparking"
version = "0.1.0"
description = ""
authors = ["Чеботарёв Виктор <powersemmi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
gunicorn = "^20.1.0"
django-environ = "^0.9.0"
cryptography = "^38.0.4"
djangorestframework = "^3.14.0"
django = "^4.1.3"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"


[tool.black]
line-length = 79
include = '\.pyi?$'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
